Encloses argument in a function's environment

This function takes a function f and a series of named arguments, and returns a closure of f which will only rely on one single argument 'param'. This is used to reduce the number of arguments passed around to likelihood or movement functions. This functionality is used internally when creating closures of custom moves in bind_moves.

bind_to_function(f, ...)

Arguments

f

The function to which arguments are bound.

...

Named arguments to bind to the function's environment.

References

Initial code comes from the partially_apply function in the 'rodeint' package (/github.com/richfitz/rodeint/).