Customise likelihood functions for outbreaker

This function is used to specify customised likelihood functions for outbreaker. Custom functions are specified as a named list or series of comma-separated, named arguments, indicating which log-likelihood component they compute. Values currently available are:

custom_likelihoods(...)

# S3 method for custom_likelihoods
print(x, ...)

Arguments

...

a named list of functions, each computing a log-likelihood component.

x

an outbreaker_config object as returned by create_config.

Value

A named list of functions with the class custom_likelihood, each implementing a customised log-likelihood components of outbreaker. Functions which are not customised will result in a NULL component.

a list of named functions

Details

genetic: the genetic likelihood; by default, the function cpp_ll_genetic is used.

timing_sampling: the likelihood of sampling times; by default, the function cpp_ll_timing_sampling is used.

timing_infections: the likelihood of infection times; by default, the function cpp_ll_timing_infections is used.

reporting: the likelihood of the reporting process; by default, the function cpp_ll_reporting is used.

contact: the likelihood of the contact tracing data; by default, the function cpp_ll_contact is used.

All log-likelihood functions should have the following arguments, in this order:

data: a list of named items containing input data as returned by outbreaker_data

param: a list of parameters with the class create_param