Small simulated outbreak

This dataset is a small (30 cases) simulated outbreak originally used to illustrate outbreaker, and used for the same purposes in outbreaker2. This list contains the following:

fake_outbreak

Format

An object of class list of length 5.

Details

$onset: A vector of integers representing dates of onset.

$sample: A vector of integers representing the dates of isolation.

$dna: A DNAbin matrix of pathogen genome sequences.

$w: A numeric vector giving the empirical distribution of the generation time.

$ances: A vector of integers indicating, for each case, the true infectors. NA represents imported cases.

Examples

names(fake_outbreak)

#> [1] "onset"  "sample" "dna"    "w"      "ances" 

fake_outbreak

#> $onset
#>  [1]  0  2  4  4  6  6  6  7  7  8  8  8  8  9  9  9  9 10 10 10 10 10 10 10 10
#> [26] 10 10 10 11 11
#> 
#> $sample
#>  [1]  3  5  6  6  7  9  8  9  9  9 11 10 10 10 10 11 11 12 11 13 12 13 11 12 11
#> [26] 11 13 12 14 14
#> 
#> $dna
#> 30 DNA sequences in binary format stored in a matrix.
#> 
#> All sequences of same length: 10000 
#> 
#> Labels:
#> ...
#> 
#> Base composition:
#>     a     c     g     t 
#> 0.251 0.242 0.251 0.256 
#> 
#> $w
#> [1] 0.04255319 0.21276596 0.42553191 0.31914894
#> 
#> $ances
#>  [1] NA  1  2 NA  3  4  4  5  6  6  7  8  9  5  5  7  7  8  9 10 11 11 13 13 13
#> [26] 17 17 NA 10 13
#>