outbreaker2: main function for reconstructing disease outbreaks

The function outbreaker is the main function of the package. It runs processes various inputs (data, configuration settings, custom priors, likelihoods and movement functions) and explores the space of plausible transmission trees of a densely sampled outbreaks.

outbreaker(data = outbreaker_data(), config = create_config(),
  priors = custom_priors(), likelihoods = custom_likelihoods(),
  moves = custom_moves())

Arguments

data

a list of named items containing input data as returned by outbreaker_data

config

a set of settings as returned by create_config

priors

a set of log-prior functions as returned by custom_priors

likelihoods

a set of log-likelihood functions as returned by custom_likelihoods

moves

a set of movement functions as returned by custom_moves

Details

The emphasis of 'outbreaker2' is on modularity, which enables customisation of priors, likelihoods and even movements of parameters and augmented data by the user. This the dedicated vignette on this topic vignette("outbreaker2_custom").

This package is in testing state. Please email the authors if you plan on using it.

References

Jombart T, Cori A, Didelot X, Cauchemez S, Fraser C and Ferguson N (2014). Bayesian reconstruction of disease outbreaks by combining epidemiologic and genomic data. PLoS Computational Biology.

See also

outbreaker_data to process input data, and create_config to process/set up parameters

outbreaker_data: function to process input data

create_config: function to create default and customise configuration settings

custom_priors: function to specify customised prior functions

custom_likelihoods: function to specify customised likelihoods functions

custom_moves: function to create default and customise movement functions

Examples

## get data
data(fake_outbreak)
dat <- fake_outbreak

## Not run: ------------------------------------
# ## run outbreaker
# out <- outbreaker(data = list(dna = dat$dna, dates = dat$onset, w_dens = dat$w),
# config = list(n_iter = 2e4, sample_every = 200))
# plot(out)
# as.data.frame(out)
# 
# ## run outbreaker, no DNA sequences
# out2 <- outbreaker(data = list(dates = dat$onset, w_dens = w),
# config = list(n_iter = 2e4, sample_every = 200))
# plot(out2)
# as.data.frame(out2)
# 
## ---------------------------------------------