Basic methods for processing outbreaker results

Several methods are defined for instances of the class outbreaker_chains, returned by outbreaker, including: print, plot

# S3 method for outbreaker_chains
print(x, n_row = 3, n_col = 8, ...)

# S3 method for outbreaker_chains
plot(x, y = "post", type = c("trace", "hist",
  "density", "alpha", "t_inf", "kappa", "network"), burnin = 0,
  min_support = 0.1, ...)

# S3 method for outbreaker_chains
summary(object, burnin = 0, ...)

Arguments

x
an outbreaker_chains object as returned by outbreaker.

n_row
the number of rows to display in head and tail; defaults to 3.

n_col
the number of columns to display; defaults to 8.

...
further arguments to be passed to other methods

y
a character string indicating which element of an outbreaker_chains object to plot

type
a character string indicating the kind of plot to be used (see details)

burnin
the number of iterations to be discarded as burnin

min_support
a number between 0 and 1 indicating the minimum support of ancestries to be plotted; only used if 'type' is 'network'

object
an outbreaker_chains object as returned by outbreaker.

Details

type indicates the type of graphic to plot:

trace to visualise MCMC traces for parameters or augmented data (plots the log-likelihood by default)

hist to plot histograms of quantitative values

density to plot kernel density estimations of quantitative values

alpha to visualise the posterior frequency of ancestries

network to visualise the transmission tree; note that this opens up an interactive plot and requires a web browser with Javascript enabled; the argument `min_support` is useful to select only the most supported ancestries and avoid displaying too many links

kappa to visualise the distributions generations between cases and their ancestor/infector

See also

See introduction vignette for detailed examples on how to visualise outbreaker_chains objects.