outbreaker2: a platform for disease outbreak reconstruction

This package provides a statistical platform for reconstructing transmission trees ('who infects whom') in densely sampled disease outbreaks. It reimplements, as a particular case, the model of 'outbreaker' (see references). 'outbreaker2' extends and replaces 'outbreaker'.

Details

The emphasis of 'outbreaker2' is on modularity, which enables customisation of priors, likelihoods and even movements of parameters and augmented data by the user. This the dedicated vignette on this topic vignette("outbreaker2_custom").

The main functions of the package are:

outbreaker: main function to run analyses

outbreaker_data: function to process input data

create_config: function to create default and customise configuration settings

custom_priors: function to specify customised prior functions

custom_likelihoods: function to specify customised likelihoods functions

custom_moves: function to create default and customise movement functions

This package is in testing state. Please email the authors if you plan on using it.