Customise samplers for outbreaker

This function is used to specify customised movement functions (a.k.a. samplers) for outbreaker. Custom functions are specified as a named list or series of comma-separated, named arguments, indicating which type of movement they implement. Values currently available are:

custom_moves(...)

# S3 method for outbreaker_moves
print(x, ...)

Arguments

...
A list or a series of named, comma-separated functions implementing movements of parameters or augmented data.

x
an outbreaker_moves object as returned by create_moves.

Value

A list of movement functions with a single argument 'param', with class outbreaker_moves.

Details

mu: movement of the mutation rate; by default, the function cpp_move_mu is used.

pi: movement of the reporting probability; by default, the function cpp_move_pi is used.

eps: movement of the contact reporting coverage; by default, the function cpp_move_eps is used.

lambda: the movement of the non-infectious contact rate; the function cpp_move_lambda is used.

alpha: movement of the transmission tree, by randomly proposing infectors in the pool of cases infected before; by default, the function cpp_move_alpha is used.

swap_cases: movement of the transmission tree, by swapping infectors and infected cases; by default, the function cpp_move_swap_cases is used.

t_inf: movement of the date of infection; by default, the function cpp_move_t_inf is used.

kappa: movement of the number generations between cases; by default, the function cpp_move_kappa is used.

Movement functions must have an argument param, which is a list of parameters and augmented data of the class create_param. Each movement function will be enclosed with its other arguments, so that the resulting function will have a single argument 'param'. For non-standard movements (i.e. none of the names specified above), the closure will contain:

data: a list of named items containing input data as returned by outbreaker_data

config: a list of named items containing input data as returned by create_config

likelihoods: a list of named custom likelihood functions as returned by custom_likelihoods

priors: a list of named custom prior functions as returned by custom_priors

See also

See customization vignette for detailed examples on how to customise movement functions.