Basic methods for processing outbreaker results

Several methods are defined for instances of the class outbreaker_chains, returned by outbreaker, including: print, plot, summary

# S3 method for outbreaker_chains
print(x, n_row = 3, n_col = 8, ...)

# S3 method for outbreaker_chains
plot(
  x,
  y = "post",
  type = c("trace", "hist", "density", "alpha", "t_inf", "kappa", "network"),
  group = NULL,
  burnin = 0,
  min_support = 0.1,
  labels = NULL,
  ...
)

# S3 method for outbreaker_chains
summary(object, burnin = 0, method = c("mpa", "decycle"), ...)

Arguments

x
an outbreaker_chains object as returned by outbreaker.

n_row
the number of rows to display in head and tail; defaults to 3.

n_col
the number of columns to display; defaults to 8.

...
further arguments to be passed to other methods

y
a character string indicating which element of an outbreaker_chains object to plot

type
a character string indicating the kind of plot to be used (see details)

group
a vector of character strings indicating the parameters to display, or "all" to display all global parameters (non node-specific parameters).

burnin
the number of iterations to be discarded as burnin

min_support
a number between 0 and 1 indicating the minimum support of ancestries to be plotted; only used if 'type' is 'network'

labels
a vector of length N indicating the case labels (must be provided in the same order used for dates of symptom onset)

object
an outbreaker_chains object as returned by outbreaker.

method
the method used to determine consensus ancestries. 'mpa' (maximum posterior ancestry) simply returns the posterior ancestry with the highest posterior support for each case, even if this includes cycles. 'decycle' will return the maximum posterior ancestry, except when cycles are detected, in which case the link in the cycle with the lowest support is pruned and the tree recalculated.

Details

type indicates the type of graphic to plot:

trace to visualise MCMC traces for parameters or augmented data (plots the log-likelihood by default)

hist to plot histograms of quantitative values

density to plot kernel density estimations of quantitative values

alpha to visualise the posterior frequency of ancestries

network to visualise the transmission tree; note that this opens up an interactive plot and requires a web browser with Javascript enabled; the argument `min_support` is useful to select only the most supported ancestries and avoid displaying too many links

kappa to visualise the distributions generations between cases and their ancestor/infector

See also

Author

Thibaut Jombart (thibautjombart@gmail.com)