Simulate contact data from a transmission tree

This function simulates contact data from a transmission tree. The model assumes that all transmission pairs have experienced contact, and that there is no false-positive reporting of contacts. The probability of contact occuring between a non-transmision pair is given by the parameter lambda. The probability of reporting a contact (transmission pair or not) is given by the parameters eps.

sim_ctd(tTree, eps, lambda)

Arguments

tTree
a dataframe or matrix of two columns, with each row providing the ids (numerical or as characters) of a transmission pair

eps
the contact reporting coverage, defined as the probability of reporting a contact (transmission pair or not)

lambda
the non-infectious contact rate, defined as the probability of contact between a non-transmission pair.

Examples

## load data
x <- fake_outbreak
tTree <- data.frame(from = x$ances, to = seq_along(x$ances))

## simulate contact data with coverage of 80% and 10% probability of
## contact between non-transmission pairs
ctd <- outbreaker2:::sim_ctd(tTree, eps = 0.8, lambda = 0.1)

## inspect contact data
head(ctd)

#>    i  j
#> 1  7 16
#> 2  9 13
#> 3 13 25
#> 4 10 29
#> 5 13 30
#> 6  4  6